# src/internal/docs/swagger.yaml
definitions:
  resource.Order:
    properties:
      created_at:
        type: integer
      id:
        type: integer
      total_amount:
        type: integer
    type: object
  resource.Status:
    properties:
      http_code:
        type: integer
    type: object
  response.Meta:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
  response.Response:
    properties:
      data: { }
      meta:
        $ref: '#/definitions/response.Meta'
    type: object
info:
  contact: { }
  title: Sample internal API
  version: 1.0.0
paths:
  /v1/orders:
    post:
      consumes:
        - application/json
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: '#/definitions/response.Response'
              - properties:
                  data:
                    $ref: '#/definitions/resource.Order'
                type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
        - BasicAuth: [ ]
      summary: Create order
      tags:
        - OrderController
  /v1/orders/{id}:
    get:
      consumes:
        - application/json
      parameters:
        - description: order id
          in: path
          name: id
          required: true
          type: integer
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: '#/definitions/response.Response'
              - properties:
                  data:
                    $ref: '#/definitions/resource.Order'
                type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
        - BasicAuth: [ ]
      summary: Get order by ID
      tags:
        - OrderController
  /v1/statuses/{code}:
    get:
      consumes:
        - application/json
      parameters:
        - description: status code
          in: path
          name: code
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
              - $ref: '#/definitions/response.Response'
              - properties:
                  data:
                    $ref: '#/definitions/resource.Status'
                type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      security:
        - BasicAuth: [ ]
      summary: API return status code
      tags:
        - StatusController
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"